@ThomasWaldmann 2015-08-13
Not sure whether I should / would accept this code.
While it's doing the right thing, it is kind of a fraction what the (still experimental) code in multithreading branch does (see my repo). So maybe rather that code should be refined further / tested more to solve a lot of scheduling problems, not just one.