@sourcejedi 2015-08-13
I see your point :).  This is clearly incremental.  E.g. I don't know whether the code would get refactored when expanding the use of threading, so all the threading code can use the same primitives and style.

I do think the concept is right.  E.g. you should still do this if you were writing multiple segment files in parallel, you shouldn't just assume their blocking IO will be fortuitously out-of-sync.