@sourcejedi 2015-08-13
Sorry, I meant to also say I will look at your threading code.