@ThomasWaldmann 2015-09-07
for master branch (not multithreading), could we get a similar speedup with less code change by going from 5MB to e.g. 50MB?