@ThomasWaldmann 2015-10-02
Just as a side note:

I am currently doing a large-scale backup for testing. The fsync somehow slows it down unnecessarily.

So to speed it up, I just removed the fsync in my local version, it's now 2x faster.