@ThomasWaldmann 2016-03-18
I am rejecting this for now and for master.

Multithreading code should go into multithreading branch (if not already implemented there) and will need a lot of testing before it gets merged into master.